refactor(blorktools): download asset button
