+++ packages/blorktools/src/asset_debugger/panels/asset-panel/asset-panel.css
@@ -69,8 +69,6 @@
     box-sizing: border-box;
 }
 
-
-
 .metadata-header {
     font-size: 1em;
     color: #aaa;
@@ -139,8 +137,6 @@
 }
 
 /* Rig Panel Styles */
-
-
 .rig-controls-section {
     display: flex;
     flex-direction: column;
@@ -150,14 +146,49 @@
     border-radius: 4px;
 }
 
-
-
 .rig-constraint-container,
 .rig-lock-container {
     display: flex;
     align-items: center;
     margin-top: 8px;
     gap: 10px;
+}
+
+/* Download Asset Section - Always visible at bottom */
+.asset-download-section {
+    margin-top: 20px;
+    padding: 15px 0;
+    border-top: 1px solid #333;
+    text-align: center;
+}
+
+.download-asset-button {
+    background-color: transparent;
+    color: #4CAF50;
+    border: 1px solid #4CAF50;
+    padding: 8px 16px;
+    font-size: 14px;
+    font-weight: normal;
+    font-family: inherit;
+    border-radius: 0;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    min-width: 160px;
+}
+
+.download-asset-button:hover {
+    background-color: rgba(76, 175, 80, 0.2);
+}
+
+.download-asset-button:disabled {
+    background-color: transparent;
+    color: #666;
+    border-color: #666;
+    cursor: not-allowed;
+}
+
+.download-asset-button:disabled:hover {
+    background-color: transparent;
 }
 
 /* Mesh Settings Modal Styles */
