+++ packages/blorktools/src/asset_debugger/panels/asset-panel/asset-panel.html
@@ -11,9 +11,6 @@
             <div id="mesh-visibility-container">
                 <h4 class="subsection-header">Mesh Visibility</h4>
                 <div id="mesh-groups" class="mesh-groups-container"></div>
-                <div class="button-container" style="margin-top: 20px; text-align: center;">
-                    <button id="download-asset-btn" class="theme-toggle">Download Asset</button>
-                </div>
             </div>
         </div>
     </div>
@@ -118,4 +115,9 @@
             </div>
         </div>
     </div>
-</div> 
+
+    <!-- Download Asset Button - Always visible at bottom -->
+    <div class="asset-download-section">
+        <button id="download-asset-btn" class="download-asset-button">Download Asset</button>
+    </div>
+</div>
